@@ -3,11 +3,23 @@
   "version": "0.0.1",
   "private": true,
   "scripts": {
-    "start": "node_modules/react-native/packager/packager.sh"
+    "start": "node node_modules/react-native/local-cli/cli.js start",
+    "test": "jest"
   },
   "dependencies": {
-    "react": "15.4.1",
-    "react-native": "0.40.0",
-    "react-native-vector-icons": "file:../../"
+    "prop-types": "^15.6.2",
+    "react": "16.5.0",
+    "react-native": "0.57.1",
+    "react-native-vector-icons": "*"
+  },
+  "devDependencies": {
+    "@babel/runtime": "^7.1.2",
+    "babel-jest": "23.6.0",
+    "jest": "23.6.0",
+    "metro-react-native-babel-preset": "0.47.0",
+    "react-test-renderer": "16.5.0"
+  },
+  "jest": {
+    "preset": "react-native"
   }
 }
