@@ -1,10 +1,12 @@
+/* eslint react/destructuring-assignment: 0 */
+/* eslint react/no-multi-comp: 0 */
+/* eslint import/no-named-as-default-member: 0 */
 import React, { PureComponent } from 'react';
+import PropTypes from 'prop-types';
 import {
-  AppRegistry,
   StyleSheet,
   Text,
   View,
-  Image,
   TabBarIOS,
   NavigatorIOS,
   TouchableOpacity,
@@ -39,13 +41,22 @@ const styles = StyleSheet.create({
 });
 
 class ColoredView extends PureComponent {
-  componentWillMount() {
+  static propTypes = {
+    color: PropTypes.string.isRequired,
+    pageText: PropTypes.string.isRequired,
+    navigator: PropTypes.shape({
+      push: PropTypes.func.isRequired,
+      pop: PropTypes.func.isRequired,
+    }).isRequired,
+  };
+
+  componentDidMount() {
     Icon.getImageSource('md-arrow-back', 30).then(source =>
       this.setState({ backIcon: source })
     );
   }
 
-  _navigateToSubview() {
+  navigateToSubview = () => {
     this.props.navigator.push({
       component: ColoredView,
       title: this.props.pageText,
@@ -53,13 +64,13 @@ class ColoredView extends PureComponent {
       onLeftButtonPress: () => this.props.navigator.pop(),
       passProps: this.props,
     });
-  }
+  };
 
   render() {
     return (
       <View style={[styles.tabContent, { backgroundColor: this.props.color }]}>
         <Text style={styles.tabText}>{this.props.pageText}</Text>
-        <TouchableOpacity onPress={() => this._navigateToSubview()}>
+        <TouchableOpacity onPress={this.navigateToSubview}>
           <View style={styles.button}>
             <Text style={styles.buttonText}>Tap Me</Text>
           </View>
@@ -69,23 +80,19 @@ class ColoredView extends PureComponent {
   }
 }
 
-class TabBarExample extends PureComponent {
-  constructor(props) {
-    super(props);
+export default class TabBarExample extends PureComponent {
+  state = {
+    selectedTab: 'home',
+  };
 
-    this.state = {
-      selectedTab: 'home',
-    };
-  }
-
-  componentWillMount() {
+  componentDidMount() {
     // https://github.com/facebook/react-native/issues/1403 prevents this to work for initial load
     Icon.getImageSource('ios-settings', 30).then(source =>
       this.setState({ gearIcon: source })
     );
   }
 
-  _renderContent(color, pageText) {
+  renderContent(color, pageText) {
     if (!this.state.gearIcon) {
       return false;
     }
@@ -108,8 +115,7 @@ class TabBarExample extends PureComponent {
       <TabBarIOS tintColor="black" barTintColor="#3abeff">
         <Icon.TabBarItemIOS
           title="Home"
-          iconName="ios-home-outline"
-          selectedIconName="ios-home"
+          iconName="ios-home"
           selected={this.state.selectedTab === 'home'}
           onPress={() => {
             this.setState({
@@ -117,25 +123,24 @@ class TabBarExample extends PureComponent {
             });
           }}
         >
-          {this._renderContent('#414A8C', 'Home')}
+          {this.renderContent('#414A8C', 'Home')}
         </Icon.TabBarItemIOS>
         <Icon.TabBarItemIOS
-          title="Profile"
-          iconName="ios-person-outline"
-          selectedIconName="ios-person"
-          selected={this.state.selectedTab === 'profile'}
+          title="Outline"
+          iconName="ios-checkbox-outline"
+          selectedIconName="ios-checkbox"
+          selected={this.state.selectedTab === 'checkbox'}
           onPress={() => {
             this.setState({
-              selectedTab: 'profile',
+              selectedTab: 'checkbox',
             });
           }}
         >
-          {this._renderContent('#090', 'Profile')}
+          {this.renderContent('#090', 'Outline')}
         </Icon.TabBarItemIOS>
         <Icon.TabBarItemIOS
           title="Starred"
-          iconName="ios-star-outline"
-          selectedIconName="ios-star"
+          iconName="ios-star"
           selected={this.state.selectedTab === 'starred'}
           onPress={() => {
             this.setState({
@@ -143,27 +148,24 @@ class TabBarExample extends PureComponent {
             });
           }}
         >
-          {this._renderContent('#900', 'Starred')}
+          {this.renderContent('#900', 'Starred')}
         </Icon.TabBarItemIOS>
         <Icon.TabBarItemIOS
           title="Settings"
-          iconName="ios-settings-outline"
-          selectedIconName="ios-settings"
+          iconName="ios-settings"
           iconColor="#ffffff"
           selectedIconColor="#000099"
           selected={this.state.selectedTab === 'settings'}
-          renderAsOriginal={true}
+          renderAsOriginal
           onPress={() => {
             this.setState({
               selectedTab: 'settings',
             });
           }}
         >
-          {this._renderContent('#009', 'Settings')}
+          {this.renderContent('#009', 'Settings')}
         </Icon.TabBarItemIOS>
       </TabBarIOS>
     );
   }
 }
-
-AppRegistry.registerComponent('TabBarExample', () => TabBarExample);
